toradex_defconfig: verdin-imx8mp: fix concurrent use of dsi and native hdmi
With native HDMI being a kernel built in and the DSI adapters being
a module,  concurrent use of the two leads to a not booting kernel.

When DRM tries to bring up native HDMI it also tries to bring up the
bridge to DSI which then returns EAGAIN as the DSI to HDMI/LVDS kernel
modules cannot be loaded yet.
This results in a never ending loop rather than continuing to init and
trying again.
Compiling the bridge to DSI as a module fixes the issue.

Upstream-Status: Inappropriate [defconfig change]
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>