arm64: dts: colibri-imx8x: Move gpio-keys to som level
SODIMM_45 that is connected to "&lsio_gpio3 10" is defined in the
Colibri standard to be a wakeup pin.
Move this to the SoM level device-tree and keep it disabled by default
but do enable it again on the carrier-board.

Upstream-Status: Backport [cc900d0ffcae8dae159b762eca3092b24744a91b]
Signed-off-by: Philippe Schenker <philippe.schenker@toradex.com>