PCI: imx: Disable vpcie in case that the PCIe PLL does not lock
If the PCIe PLL locking fails we also have to disable vpcie because it
was already enabled before. Without this commit we get a warning when
the kernel tries to free the vpcie regulator.

Upstream-Status: Pending

Mainline for i.MX 8X does not support PCIe yet, this patch cannot be
upstreamed as of now.

Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>