arm64: dts: colibri-imx8x: use clock and regulator instead of gpio hogs
Replace gpio-hogs by using the hsio clock to enable the PCIe reference
clock for Wifi and vcie-supply to power on the Wifi module. With this
change the gpios will not be touched when Wifi is disabled.

Upstream-Status: Pending

Mainline for i.MX 8X does not support PCIe yet, this patch cannot be
upstreamed as of now.

Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>