usb: chipidea: remove broken suspend/resume downstream code
The suspend/resume code is broken for i.MX 8 CPUs. When waking up from
suspend it might happen that devices are not recognized anymore. To
reproduce do:
root@apalis-imx8-07107223:~# lsusb -t
/:  Bus 03.Port 1: Dev 1, Class=root_hub, Driver=ci_hdrc/1p, 480M
    |__ Port 1: Dev 2, If 0, Class=Hub, Driver=hub/3p, 480M
        |__ Port 1: Dev 4, If 0, Class=Wireless, Driver=btusb, 480M
        |__ Port 1: Dev 4, If 1, Class=Wireless, Driver=btusb, 480M
        |__ Port 2: Dev 3, If 0, Class=Hub, Driver=hub/4p, 480M
/:  Bus 02.Port 1: Dev 1, Class=root_hub, Driver=xhci-hcd/1p, 5000M
/:  Bus 01.Port 1: Dev 1, Class=root_hub, Driver=xhci-hcd/1p, 480M
root@apalis-imx8-12345678:~# rtcwake -s 1 -m mem -d rtc1
root@apalis-imx8-07107223:~# lsusb -t
/:  Bus 03.Port 1: Dev 1, Class=root_hub, Driver=ci_hdrc/1p, 480M
/:  Bus 02.Port 1: Dev 1, Class=root_hub, Driver=xhci-hcd/1p, 5000M
/:  Bus 01.Port 1: Dev 1, Class=root_hub, Driver=xhci-hcd/1p, 480M

We remove this broken code to replace it with the upstream code for
suspend and resume.

Upstream-Status: Inappropriate [other]

This is an issue of the downstream driver. We replace the broken
suspend/resume with the upstream solution.

Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>