usb: dwc3: imx8mp: add bus_early_clk
Add a bus_early_clk which is the same clock as for the usb controllers
child. This is necessary because the mentioned commit in fixes is
accessing registers which need the IMX8MP_CLK_HSI_AXI clock to be
enabled. Without this fix suspend/resume is not working properly if no
USB-devices are connected.
rtcwake -s 5 -m mem -d rtc1
It will try to wakeup after 5 seconds but hangs afterward.

Upstream-Status: Inappropriate [other]

This fix is only needed for the downstream kernel. In the upstream
kernel they added a power-domain for all hsio devices (usb, pcie, etc.)
and enable the clock there. However, backporting this solution would
probably cause more issues for now.

Fixes: cbd09f688623 ("usb: dwc3: imx8mp: Add support for setting SOC specific flags")
Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>