usb: dwc3: imx8mp: fix bus_early device tree property name
When adding the bus_early device tree property, bus_early_clk was
incorrectly used as name. This commit changes the property name to
bus_early, which is consistent with the other clocks.

Upstream-Status: Inappropriate [other]

This fix is only needed for the downstream kernel. In the upstream
kernel they added a power-domain for all hsio devices (usb, pcie, etc.)
and enable the clock there. However, backporting this solution would
probably cause more issues for now.

Fixes: 4d948676c55f("usb: dwc3: imx8mp: add bus_early_clk")
Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>