Revert "mmc: host: sdhci-esdhc-imx.c: correct the auto-tuning setting for sdio device"
Revert this commit. We will instead backport
commit c8c49a5a0b2b ("mmc: sdhci-esdhc-imx: simplify the auto tuning logic")
and commit
52e4c32baed2 ("mmc: sdhci-esdhc-imx: only enable DAT[0] and CMD line auto tuning for SDIO device")
from upstream that solve the same problem in a different way.

This way we are more close to mainline and we do not need custom properties
into the Verdin iMX8M Plus and Mini dts files.

Upstream-Status: Inappropriate [other]

This commit does not exist upstream and the same issue is already solved there.

Signed-off-by: Emanuele Ghidoli <emanuele.ghidoli@toradex.com>