usb: mux: gpio-switch: correct GPIOLIB dependency
When GPIOLIB is not enabled, for example as a kernel .config output from
OE, gpio-switch driver fail the kernel build. The problem is that the
driver include an incorrect include file for gpio functions. The change
just update the driver to use the correct header file for the gpio
consumer API that already handles GPIOLIB dependency.

Upstream-Status: Inappropriate [other]

The gpio-switch driver is downstream only.

Signed-off-by: Joao Paulo Goncalves <joao.goncalves@toradex.com>