drm/bridge: ti-sn65dsi83: Fix rmmod crash
When unloading the module the driver crashes because it is still
registered and used in the drm chain. This patch fixes it detaching and
disabling the bridge before removing the bridge.

Upstream-Status: Inappropriate [other]

The upstream driver uses devm functions and is most likely not affected
by this issue. However, kernel 5.15 does not support them yet.

Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>