arm64: dts: imx8mm-verdin: overwrite alloc-ranges property
With commit 757e132269ec ("arm64: dts: imx8mm-verdin: fix cma memory
region") we introduced a new problem for modules with more than 3GB of
RAM. The VPU driver Hantro is currently unable to cross 32bit
boundaries. However, since the cma region is located at the end of the
RAM, we have crossed the boundary for modules with more than 3GB of RAM.
The Verdin iMX8MM modules are not yet affected by this issue as there is
no module with more than 2GB of RAM. However, to be consistent with the
Verdin iMX8MP and to be future proof in case there is a module with more
RAM, we overwrite the alloc-ranges property here as well. We cannot use
the default alloc-ranges property from imx8mm.dtsi because it starts at
1GB of RAM and therefore does not work with modules that only have 1GB
of RAM. That is why we deleted this property in the first place.

Upstream-Status: Inappropriate [other]

We use the kernel configuration to define the CMA size and don't use the
linux,cma node in the device tree. However, we need to verify that the
upstream hantro driver can handle 64 bit addresses.

Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>