arm64: dts: imx8mp-verdin: overwrite alloc-ranges property
With commit 1263f108fb28 ("arm64: dts: imx8mp-verdin: fix cma memory
region") we introduced a new problem for modules with more than 3GB of
RAM. The VPU driver Hantro is currently unable to cross 32bit
boundaries. However, since the cma region is at the end of the RAM, we
have crossed the boundary for modules with more than 3GB of RAM. We set
the alloc-ranges property to the same value that is used in imx8mp.dtsi.
We do this to be consistent between Verdin iMX8MM and Verdin iMX8MP.

Upstream-Status: Inappropriate [other]

We use the kernel configuration to define the CMA size and don't use the
linux,cma node in the device tree. However, we need to verify that the
upstream hantro driver can handle 64 bit addresses.

Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>